requirements.txt
aiohttp
aiofiles
asyncio
pillow
